mod #61 : 파일 구조 변경
