+++ src/main/java/com/lokoko/domain/product/controller/ProductController.java
@@ -9,23 +9,23 @@ import static com.lokoko.domain.product.controller.enums.ResponseMessage.PRODUCT
 import static com.lokoko.domain.product.controller.enums.ResponseMessage.PRODUCT_YOUTUBE_DETAIL_SUCCESS;
 
 import com.lokoko.domain.product.controller.enums.ResponseMessage;
-import com.lokoko.domain.product.dto.CategoryNewProductResponse;
-import com.lokoko.domain.product.dto.CategoryPopularProductResponse;
-import com.lokoko.domain.product.dto.CategoryProductPageResponse;
-import com.lokoko.domain.product.dto.CrawlRequest;
-import com.lokoko.domain.product.dto.NameBrandProductResponse;
-import com.lokoko.domain.product.dto.ProductDetailResponse;
-import com.lokoko.domain.product.dto.ProductDetailYoutubeResponse;
+import com.lokoko.domain.product.dto.response.CategoryNewProductResponse;
+import com.lokoko.domain.product.dto.response.CategoryPopularProductResponse;
+import com.lokoko.domain.product.dto.response.CategoryProductPageResponse;
+import com.lokoko.domain.product.dto.request.CrawlRequest;
+import com.lokoko.domain.product.dto.response.NameBrandProductResponse;
+import com.lokoko.domain.product.dto.response.ProductDetailResponse;
+import com.lokoko.domain.product.dto.response.ProductDetailYoutubeResponse;
 import com.lokoko.domain.product.entity.enums.MiddleCategory;
 import com.lokoko.domain.product.entity.enums.SubCategory;
 import com.lokoko.domain.product.service.NewProductCrawlingService;
 import com.lokoko.domain.product.service.ProductCrawlingService;
 import com.lokoko.domain.product.service.ProductReadService;
 import com.lokoko.domain.product.service.ProductService;
-import com.lokoko.domain.review.dto.ImageReviewListResponse;
-import com.lokoko.domain.review.dto.KeywordImageReviewListResponse;
-import com.lokoko.domain.review.dto.KeywordVideoReviewListResponse;
-import com.lokoko.domain.review.dto.VideoReviewListResponse;
+import com.lokoko.domain.review.dto.response.ImageReviewListResponse;
+import com.lokoko.domain.review.dto.response.KeywordImageReviewListResponse;
+import com.lokoko.domain.review.dto.response.KeywordVideoReviewListResponse;
+import com.lokoko.domain.review.dto.response.VideoReviewListResponse;
 import com.lokoko.domain.review.exception.MissingMediaTypeException;
 import com.lokoko.domain.review.service.ReviewReadService;
 import com.lokoko.global.common.entity.MediaType;
