+++ src/main/java/com/lokoko/domain/product/dto/request/CrawlRequest.java
@@ -1,4 +1,4 @@
-package com.lokoko.domain.product.dto;
+package com.lokoko.domain.product.dto.request;
 
 import com.lokoko.domain.product.entity.enums.MainCategory;
 import com.lokoko.domain.product.entity.enums.MiddleCategory;
