+++ src/main/java/com/lokoko/domain/product/dto/response/CategoryNewProductResponse.java
@@ -1,9 +1,9 @@
-package com.lokoko.domain.product.dto;
+package com.lokoko.domain.product.dto.response;
 
 import com.lokoko.global.common.response.PageableResponse;
 import java.util.List;
 
-public record NameBrandProductResponse(
+public record CategoryNewProductResponse(
         String searchQuery,
         List<ProductResponse> products,
         PageableResponse pageInfo
