+++ src/main/java/com/lokoko/domain/product/dto/response/CategoryPopularProductResponse.java
@@ -1,9 +1,9 @@
-package com.lokoko.domain.product.dto;
+package com.lokoko.domain.product.dto.response;
 
 import com.lokoko.global.common.response.PageableResponse;
 import java.util.List;
 
-public record CategoryNewProductResponse(
+public record CategoryPopularProductResponse(
         String searchQuery,
         List<ProductResponse> products,
         PageableResponse pageInfo
