+++ src/main/java/com/lokoko/domain/product/dto/response/CategoryProductPageResponse.java
@@ -1,10 +1,13 @@
-package com.lokoko.domain.product.dto;
+package com.lokoko.domain.product.dto.response;
 
 import com.lokoko.global.common.response.PageableResponse;
 import java.util.List;
+import lombok.Builder;
 
-public record CategoryPopularProductResponse(
+@Builder
+public record CategoryProductPageResponse(
         String searchQuery,
+        String parentCategoryName,
         List<ProductResponse> products,
         PageableResponse pageInfo
 ) {
