+++ src/main/java/com/lokoko/domain/product/dto/response/NameBrandProductResponse.java
@@ -1,0 +1,11 @@
+package com.lokoko.domain.product.dto.response;
+
+import com.lokoko.global.common.response.PageableResponse;
+import java.util.List;
+
+public record NameBrandProductResponse(
+        String searchQuery,
+        List<ProductResponse> products,
+        PageableResponse pageInfo
+) {
+}
