+++ src/main/java/com/lokoko/domain/product/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package com.lokoko.domain.product.dto;
+package com.lokoko.domain.product.dto.response;
 
 
 import java.util.List;
