+++ src/main/java/com/lokoko/domain/product/service/ProductReadService.java
@@ -3,14 +3,14 @@ package com.lokoko.domain.product.service;
 
 import com.lokoko.domain.image.entity.ProductImage;
 import com.lokoko.domain.image.repository.ProductImageRepository;
-import com.lokoko.domain.product.dto.CategoryNewProductResponse;
-import com.lokoko.domain.product.dto.CategoryPopularProductResponse;
-import com.lokoko.domain.product.dto.CategoryProductPageResponse;
-import com.lokoko.domain.product.dto.ProductDetailResponse;
-import com.lokoko.domain.product.dto.ProductDetailYoutubeResponse;
-import com.lokoko.domain.product.dto.ProductOptionResponse;
-import com.lokoko.domain.product.dto.ProductResponse;
-import com.lokoko.domain.product.dto.ProductSummary;
+import com.lokoko.domain.product.dto.response.CategoryNewProductResponse;
+import com.lokoko.domain.product.dto.response.CategoryPopularProductResponse;
+import com.lokoko.domain.product.dto.response.CategoryProductPageResponse;
+import com.lokoko.domain.product.dto.response.ProductDetailResponse;
+import com.lokoko.domain.product.dto.response.ProductDetailYoutubeResponse;
+import com.lokoko.domain.product.dto.response.ProductOptionResponse;
+import com.lokoko.domain.product.dto.response.ProductResponse;
+import com.lokoko.domain.product.dto.response.ProductSummary;
 import com.lokoko.domain.product.entity.Product;
 import com.lokoko.domain.product.entity.ProductOption;
 import com.lokoko.domain.product.entity.enums.MiddleCategory;
