+++ src/main/java/com/lokoko/domain/review/dto/response/ImageReviewListResponse.java
@@ -1,4 +1,4 @@
-package com.lokoko.domain.review.dto;
+package com.lokoko.domain.review.dto.response;
 
 import com.lokoko.global.common.response.PageableResponse;
 import java.util.List;
