+++ src/main/java/com/lokoko/domain/review/repository/ReviewRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.lokoko.domain.image.entity.QReviewImage;
 import com.lokoko.domain.product.entity.QProduct;
 import com.lokoko.domain.product.entity.enums.MiddleCategory;
 import com.lokoko.domain.product.entity.enums.SubCategory;
-import com.lokoko.domain.review.dto.ImageReviewResponse;
-import com.lokoko.domain.review.dto.VideoReviewResponse;
+import com.lokoko.domain.review.dto.response.ImageReviewResponse;
+import com.lokoko.domain.review.dto.response.VideoReviewResponse;
 import com.lokoko.domain.review.entity.QReview;
 import com.lokoko.domain.video.entity.QReviewVideo;
 import com.querydsl.core.BooleanBuilder;
